+++ twitchup.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import re
 import sys
@@ -12,9 +13,18 @@ reddit = praw.Reddit(
     client_secret=os.environ['REDDIT_CLIENT_SECRET'],
     username=os.environ['REDDIT_USERNAME'],
     password=os.environ['REDDIT_PASSWORD'],
-    user_agent=f"{sys.platform}:twitchup:0.1.1 (by /u/Volcyy)"
+    user_agent=f"{sys.platform}:twitchup:0.2.0 (by /u/Volcyy)"
 )
-COMMAND_RE = re.compile(r'(?<=twitchup\()(.*?)(?=\))')
+
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s | %(name)-10s | %(levelname)-8s | %(message)s'
+)
+logging.getLogger('urllib3.connectionpool').setLevel(logging.WARN)
+logging.getLogger('prawcore').setLevel(logging.WARN)
+log = logging.getLogger('twitchup')
+
+COMMAND_RE = re.compile(r'twitchup\((\w+)\)')
 SUBREDDIT_NAME = os.environ['SUBREDDIT_NAME']
 
 
@@ -30,7 +40,7 @@ class TwitchClient:
 
         if response.headers['Ratelimit-Remaining'] == 0:
             sleep_seconds = response.headers['Ratelimit-Remaining'] - time.time()
-            print(f'sleeping for {sleep_seconds} as no requests are left in this time period')
+            log.info(f'Sleeping for {sleep_seconds} as no requests are left in this time period.')
             time.sleep(sleep_seconds)
 
         response.raise_for_status()
@@ -42,27 +52,44 @@ class TwitchClient:
 
 
 if __name__ == '__main__':
-    with open('template.md', 'r') as f:
-        sidebar = f.read()
+    sidebars = {}
+    for template in os.listdir('templates'):
+        with open(f'templates/{template}', 'r') as f:
+            subreddit_name = template.rstrip('.md')
+            sidebars[subreddit_name] = f.read()
+            log.info(f"Obtained template for /r/{subreddit_name}.")
 
     twitch = TwitchClient()
+    streams = {
+        stream_name: twitch.get_stream(stream_name)
+        for stream_name in {
+            match.group(0).lstrip('twitchup(').rstrip(')')
+            for sidebar in sidebars.values()
+            for match in COMMAND_RE.finditer(sidebar)
+        }
+    }
+    log.info(f"Loaded stream information for {len(streams)} streams.")
 
-    for match in COMMAND_RE.finditer(sidebar):
-        stream_name = match.group(0).strip('()')
-        stream_data = twitch.get_stream(stream_name)
-        if stream_data is not None:
-            link_title = 'twitch-online'
+    for subreddit_name, sidebar in sidebars.items():
+        for match in COMMAND_RE.finditer(sidebar):
+            stream_name = match.group(0).lstrip('twitchup(').rstrip(')')
+            if streams[stream_name] is not None:
+                link_title = 'twitch-online'
+            else:
+                link_title = 'twitch-offline'
+
+            markdown_link = f"[{stream_name}](https://twitch.tv/{stream_name} '{link_title}')"
+            sidebar = sidebar.replace(f'twitchup({stream_name})', markdown_link)
+            log.debug(f"Added Twitch link on /r/{subreddit_name} for https://twitch.tv/{stream_name}.")
+
+        mod_relationship = reddit.subreddit(subreddit_name).mod
+        old_description = mod_relationship.settings()['description']
+
+        # Only update the sidebar if we made any actual changes.
+        if sidebar != old_description:
+            mod_relationship.update(description=sidebar)
+            log.info(f"Updated sidebar on {subreddit_name} with new stream data.")
         else:
-            link_title = 'twitch-offline'
+            log.info(f"Omitting update on {subreddit_name} as no changes would be done.")
 
-        markdown_link = f"[{stream_name}](https://twitch.tv/{stream_name} '{link_title}')"
-        sidebar = sidebar.replace(f'twitchup({stream_name})', markdown_link)
-
-    mod_relationship = reddit.subreddit(SUBREDDIT_NAME).mod
-    old_description = mod_relationship.settings()['description']
-
-    # Only update the sidebar if we made any actual changes.
-    if sidebar != old_description:
-        mod_relationship.update(description=sidebar)
-    else:
-        print('not updating sidebar as there is no difference to the current one')
+    log.info("Done.")