+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project aims to adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.2.0] - 2018-09-15
 ### Added
 - The subreddit sidebar is no longer updated if no changes would be made by an update.
+- Now supports multiple templates.
+- Added a `Dockerfile`.
+- Added logging of noteworthy events.
 
 ### Changed
 - The project now uses `Pipfile` instead of `pyproject.toml`.